refactoring

+++ src/components/Dashboard/DashboardLogic.jsx
@@ -1,7 +1,7 @@
-import React from "react";
-import AddFoodItemButton from "./AddFoodItemButton";
-import MoveItemButton from "./MoveItemButton";
-import { database } from "../FirebaseConfig";
+
+import { useLists } from "../../contexts/ListsContext";
+import { useAuth } from "../../contexts/AuthContext";
+import { database } from "../../FirebaseConfig";
 import {
   collection,
   getDocs,
@@ -10,33 +10,52 @@ import {
   doc,
   writeBatch,
 } from "firebase/firestore";
-import ManageButton from "./ManageButton";
-import { useLists } from "../contexts/ListsContext";
 
-export default function DashboardLeft({
-  setManageOverlay,
-  manageOverlay
-}) {
-  const {activeListSize,
-    inactiveListSize,
-    maxListId,
-    lists,
-    setLists,
-    selectedItems,
-    setSelectedItems,
-    listsDragged,
-    setListsDragged,} = useLists()
+export default function DashboardLogic({}) {
+    const {activeListSize,
+        inactiveListSize,
+        maxListId,
+        setLists,
+        lists,
+        selectedItems,
+        setSelectedItems,
+        listsDragged,
+        setListsDragged,} = useLists()
 
-  const deletedSelected = () => {
+  const { currentUser } = useAuth();
+
+//  console.log(lists)
+
+const deleteFromDB = async (itemsToDelete) => {
+  try {
+    const itemsRef = collection(database, "foodItems");
+    const batch = writeBatch(database);
+    for (const item of itemsToDelete) {
+      const itemQuery = query(itemsRef, where("id", "==", item.id));
+      const itemDocs = await getDocs(itemQuery);
+      if (!itemDocs.empty) {
+        const itemDoc = itemDocs.docs[0];
+        const itemRef = doc(itemsRef, itemDoc.id);
+
+        batch.delete(itemRef);
+      }
+    }
+    await batch.commit();
+    console.log("Items deleted successfully from the database");
+  } catch (error) {
+    console.error("Error deleting items from the database:", error);
+  }
+};
+
+  const deleteSelected = () => {
     setLists((oldLists) => {
       const toDelete = [];
       const updatedLists = oldLists.map((list) => {
-        const {items} = list;
-        let newListItems = [...items]
+        const { items } = list;
+        let newListItems = [...items];
         newListItems.forEach((item) => {
           if (selectedItems.includes(item)) {
-            // console.log("active?", item.isActive)
-            if (item.isActive) activeListSize.current -= 1
+            if (item.isActive) activeListSize.current -= 1;
             else inactiveListSize.current -= 1;
             toDelete.push(item);
           }
@@ -44,12 +63,15 @@ export default function DashboardLeft({
         newListItems = newListItems.filter(
           (item) => !selectedItems.includes(item)
         );
-        return {...list, items: newListItems};
+        return { ...list, items: newListItems };
       });
-      deleteFromDB(toDelete);
+      if (currentUser) {
+        deleteFromDB(toDelete);
+      }
       return updatedLists;
     });
     setSelectedItems([]);
+    // console.log(activeListSize.current)
   };
 
   const moveToInactiveListInverse = () => {
@@ -58,7 +80,6 @@ export default function DashboardLeft({
       const updatedLists = oldLists.map((list) => {
         const listName = list.name;
         let listItems = [...list.items];
-        const color = list.color
         if (listName === "inactive") {
           toInactive = toInactive.map((item) => {
             activeListSize.current -= 1;
@@ -69,8 +90,10 @@ export default function DashboardLeft({
             };
           });
           toInactive.forEach((item) => listItems.push(item));
-          modifyActiveDBStatus(toInactive);
-          return {...list, name: listName, items: listItems};
+          if (currentUser) {
+            modifyActiveDBStatus(toInactive);
+          }
+          return { ...list, name: listName, items: listItems };
         } else {
           listItems.forEach((item) => {
             if (!selectedItems.includes(item)) {
@@ -80,7 +103,7 @@ export default function DashboardLeft({
               selectedItems.includes(item)
             );
           });
-          return {...list, items: listItems};
+          return { ...list, items: listItems };
         }
       });
       return updatedLists;
@@ -95,16 +118,11 @@ export default function DashboardLeft({
         const listName = list.name;
         let listItems = [...list.items];
         if (listName === "inactive") {
-          // Store inactive items to make sure items arent duplcated later
-          
           const inactiveItems = listItems.map((item) => {
-            // console.log('item', item.id)
-            return item
+            return item;
           });
-          // console.log(inactiveItemIds)
           selectedItems.forEach((item) => {
             if (!inactiveItems.includes(item)) {
-              // console.log(item)
               activeListSize.current -= 1;
               inactiveListSize.current += 1;
               listItems.push(item);
@@ -120,9 +138,11 @@ export default function DashboardLeft({
               )
           );
         }
-        return {...list, items: listItems};
+        return { ...list, items: listItems };
       });
-      modifyActiveDBStatus(toDatabase);
+      if (currentUser) {
+        modifyActiveDBStatus(toDatabase);
+      }
       return updatedLists;
     });
     setSelectedItems([]);
@@ -151,9 +171,11 @@ export default function DashboardLeft({
             }
           });
         }
-        return {...list, items: listItems};
+        return { ...list, items: listItems };
       });
-      modifyActiveDBStatus(toDatabase);
+      if (currentUser) {
+        modifyActiveDBStatus(toDatabase);
+      }
       return updatedLists;
     });
     setSelectedItems([]);
@@ -186,26 +208,6 @@ export default function DashboardLeft({
     }
   };
 
-  const deleteFromDB = async (itemsToDelete) => {
-    try {
-      const itemsRef = collection(database, "foodItems");
-      const batch = writeBatch(database);
-      for (const item of itemsToDelete) {
-        const itemQuery = query(itemsRef, where("id", "==", item.id));
-        const itemDocs = await getDocs(itemQuery);
-        if (!itemDocs.empty) {
-          const itemDoc = itemDocs.docs[0];
-          const itemRef = doc(itemsRef, itemDoc.id);
-
-          batch.delete(itemRef);
-        }
-      }
-      await batch.commit();
-      console.log("Items deleted successfully from the database");
-    } catch (error) {
-      console.error("Error deleting items from the database:", error);
-    }
-  };
   const addList= (listName) => {
     maxListId.current += 1
     const newList = {name: listName, color: maxListId.current, id: maxListId.current, items: []}
@@ -217,41 +219,11 @@ export default function DashboardLeft({
     setListsDragged(!listsDragged);
   }
 
-  return (
-    <div className="left">
-      <AddFoodItemButton
-        setLists={setLists}
-        lists={lists}
-        activeListSize={activeListSize}
-        addList={addList}
-        maxListId={maxListId}
-      />
-      <MoveItemButton
-        checkedItems={selectedItems}
-        actionFunction={moveToInactiveList}
-        buttonName="Set Inactive"
-      />
-      <MoveItemButton
-        checkedItems={selectedItems}
-        actionFunction={moveToActiveList}
-        buttonName="Set Active"
-      />
-      <MoveItemButton
-        checkedItems={selectedItems}
-        actionFunction={moveToInactiveListInverse}
-        buttonName="Deactivate Unselected"
-      />
-      <MoveItemButton
-        checkedItems={selectedItems}
-        actionFunction={deletedSelected}
-        buttonName="Delete Selected"
-      />
-      <ManageButton
-        setManageOverlay={setManageOverlay}
-        manageOverlay={manageOverlay}
-        lists={lists}
-        setLists={setLists}
-        />
-    </div>
-  )
+  return {
+    deleteSelected,
+    moveToInactiveListInverse,
+    moveToInactiveList,
+    moveToActiveList,
+    addList,
+  };
 }
